Separated OPTICS and extraction functions
